Code cleanup.

+++ engine/engine-base/src/main/java/server/MapleStatEffect.java
@@ -335,7 +335,7 @@ public class MapleStatEffect {
             if (projectile == null) {
                return false;
             } else {
-               MapleInventoryManipulator.removeFromSlot(applyto.getClient(), MapleInventoryType.USE, projectile.position(), (short) projectileConsume, false, true);
+               MapleInventoryManipulator.removeFromSlot(applyto.getClient(), MapleInventoryType.USE, projectile.position(), projectileConsume, false, true);
             }
          } finally {
             use.unlockInventory();
